♻️ refactor(monitoring): extract shared monitoring UI
Move summary cards, pagination controls, status badges, and recent-pattern bars into shared monitoring components.

This keeps MonitoringCenterPage focused on orchestration while preserving existing rendering; validated with type-check, targeted tests, and lint.
