+++ src/features/monitoring/components/MonitoringShared.tsx
@@ -1,0 +1,154 @@
+import type { ReactNode } from 'react';
+import type { TFunction } from 'i18next';
+import { Button } from '@/components/ui/Button';
+import { Card } from '@/components/ui/Card';
+import { Select } from '@/components/ui/Select';
+import type { MonitoringStatusTone } from '@/features/monitoring/hooks/useMonitoringData';
+import styles from '../MonitoringCenterPage.module.scss';
+
+export type SummaryCardProps = {
+  label: string;
+  value: string;
+  meta: string;
+  tone?: MonitoringStatusTone;
+  variant?: 'primary' | 'secondary';
+};
+
+type PaginationControlsProps = {
+  count: number;
+  currentPage: number;
+  totalPages: number;
+  startItem: number;
+  endItem: number;
+  pageSize: number;
+  pageSizeOptions: readonly number[];
+  onPageChange: (page: number) => void;
+  onPageSizeChange: (pageSize: number) => void;
+  t: TFunction;
+};
+
+const parsePageSize = (value: string, fallback: number) => {
+  const parsed = Number.parseInt(value, 10);
+  return Number.isFinite(parsed) && parsed > 0 ? parsed : fallback;
+};
+
+export function SummaryCard({ label, value, meta, tone, variant = 'primary' }: SummaryCardProps) {
+  const cardClassName = [
+    styles.summaryCard,
+    variant === 'secondary' ? styles.summaryCardSecondary : styles.summaryCardPrimary,
+  ]
+    .filter(Boolean)
+    .join(' ');
+
+  return (
+    <Card className={cardClassName}>
+      <span className={styles.summaryLabel}>{label}</span>
+      <strong className={`${styles.summaryValue} ${tone ? styles[`tone${tone}`] : ''}`}>
+        {value}
+      </strong>
+      <span className={styles.summaryMeta}>{meta}</span>
+    </Card>
+  );
+}
+
+export function PaginationControls({
+  count,
+  currentPage,
+  totalPages,
+  startItem,
+  endItem,
+  pageSize,
+  pageSizeOptions,
+  onPageChange,
+  onPageSizeChange,
+  t,
+}: PaginationControlsProps) {
+  if (count === 0) return null;
+
+  return (
+    <div className={styles.paginationBar}>
+      <div className={styles.paginationInfo}>
+        {t('monitoring.pagination_info', {
+          current: currentPage,
+          total: totalPages,
+          start: startItem,
+          end: endItem,
+          count,
+        })}
+      </div>
+      <div className={styles.paginationControls}>
+        <div className={styles.pageSizeField}>
+          <span>{t('monitoring.page_size_label')}</span>
+          <Select
+            className={styles.pageSizeSelect}
+            triggerClassName={styles.pageSizeSelectTrigger}
+            value={String(pageSize)}
+            options={pageSizeOptions.map((size) => ({
+              value: String(size),
+              label: t('monitoring.page_size_option', { count: size }),
+            }))}
+            onChange={(value) => onPageSizeChange(parsePageSize(value, pageSize))}
+            ariaLabel={t('monitoring.page_size_label')}
+            fullWidth={false}
+          />
+        </div>
+        <Button
+          variant="secondary"
+          size="sm"
+          onClick={() => onPageChange(Math.max(1, currentPage - 1))}
+          disabled={currentPage <= 1}
+        >
+          {t('monitoring.pagination_prev')}
+        </Button>
+        <Button
+          variant="secondary"
+          size="sm"
+          onClick={() => onPageChange(Math.min(totalPages, currentPage + 1))}
+          disabled={currentPage >= totalPages}
+        >
+          {t('monitoring.pagination_next')}
+        </Button>
+      </div>
+    </div>
+  );
+}
+
+export function StatusBadge({
+  tone,
+  children,
+}: {
+  tone: MonitoringStatusTone;
+  children: ReactNode;
+}) {
+  return <span className={`${styles.statusBadge} ${styles[`tone${tone}`]}`}>{children}</span>;
+}
+
+export function RecentPattern({
+  pattern,
+  variant = 'default',
+}: {
+  pattern: boolean[];
+  variant?: 'default' | 'plain';
+}) {
+  const normalized = pattern.length > 0 ? pattern : Array.from({ length: 10 }, () => true);
+  const containerClassName = [
+    styles.patternBars,
+    variant === 'plain' ? styles.patternBarsPlain : '',
+  ]
+    .filter(Boolean)
+    .join(' ');
+  const barClassName = [styles.patternBar, variant === 'plain' ? styles.patternBarPlain : '']
+    .filter(Boolean)
+    .join(' ');
+
+  return (
+    <div className={containerClassName} aria-hidden="true">
+      {normalized.map((item, index) => (
+        <span
+          key={`${index}-${item ? 'success' : 'failed'}`}
+          className={`${barClassName} ${item ? styles.patternSuccess : styles.patternFailed}`}
+        />
+      ))}
+    </div>
+  );
+}
